batch: Isolate replacement group verification
Individual replacement locations have dedicated proof helpers. Related
child replacements can also share evidence for their complete historical
parent.

The group proof and its temporary occurrence index still live beside
per-unit dispatch, obscuring which evidence may be reused by siblings.

This commit moves whole-parent and partial-child verification into a group
planning module while preserving the existing proof bodies.
